requirement.txt
groq
uvicorn
fastapi
pypdf
numpy
chromadb
sentence-transformers
torch
python-dotenv
pydantic
typing-extensions
tqdm
scikit-learn